requirements.txt
miniflux
openai
tiktoken
markdownify
markdown
PyYAML
flask
feedgen
schedule
ratelimit
google-genai
